requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.0.0
redis==2.10.6
requests==2.18.4
urllib3==1.22
Werkzeug==0.14.1
honeycomb-beeline==2.11.4
